added createRangeControl component
